feat(#2589): allow to set or override group icons in the instance sidebar (#2608)
